# frntEnd/openapi.yaml
openapi: 3.1.0
info:
  title: AI Workflow Builder API
  version: 1.0.0
  description: |
    REST API for the AI Workflow Builder platform.

    ## Authentication
    All endpoints (except `/auth/*` and `/health`) require a valid JWT Bearer token
    or an API Key in the `X-API-Key` header.

    ## Tenant Isolation
    Every request is scoped to a tenant. The tenant is resolved from the JWT claim
    (`tenant_id`) or from the API key row. Cross-tenant access is impossible by design.

    ## Error Envelope
    All error responses use the standard envelope below:
    ```json
    {
      "error": {
        "code": "WORKFLOW_NOT_FOUND",
        "message": "Workflow wf_abc123 not found.",
        "request_id": "req_01J..."
      }
    }
    ```

  contact:
    name: Platform Engineering
  license:
    name: Proprietary
servers:
  - url: https://api.workflowplatform.io/v1
    description: Production (us-east-1)
  - url: https://api.eu.workflowplatform.io/v1
    description: EU Production (eu-west-1 — v1.2+)
  - url: http://localhost:8000/v1
    description: Local development

# ─────────────────────────────────────────────────────────────────────────────
# Security Schemes
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256-signed JWT. Obtain via `/auth/login` or `/auth/token/refresh`.

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Tenant API key (prefix `wfk_`). Created in the dashboard.
        Scopes are checked per-endpoint.

  # ───────────────────────────────────────────────────────────────────────────
  # Reusable Schemas
  # ───────────────────────────────────────────────────────────────────────────
  schemas:
...
    ErrorDetail:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          example: WORKFLOW_NOT_FOUND
        message:
          type: string
          example: "Workflow wf_abc123 not found."
        request_id:
          type: string
          example: req_01JRXYZ
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'

    # ── Pagination ─────────────────────────────────────────────────────────────
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        has_next:
          type: boolean

    # ── Auth ───────────────────────────────────────────────────────────────────
...
    ScheduleCreate:
      type: object
      required: [cron_expression, timezone]
      properties:
        cron_expression:
          type: string
          description: Standard 5-field cron expression.
          example: "0 9 * * MON-FRI"
        timezone:
          type: string
          description: IANA timezone identifier.
          example: "America/New_York"
        trigger_input:
          type: object
          description: Static payload to inject at each fire.
        is_active:
          type: boolean
          default: true
    Schedule:
      allOf:
        - $ref: '#/components/schemas/ScheduleCreate'
        - type: object
          required: [schedule_id, workflow_id, next_fire_at, created_at]
          properties:
            schedule_id:
              type: string
            workflow_id:
              type: string
            next_fire_at:
              type: string
              format: date-time
            last_fired_at:
              type: string
              format: date-time
              nullable: true
            created_at:
              type: string
              format: date-time

    # ── Webhooks ───────────────────────────────────────────────────────────────
    WebhookCreate:
      type: object
      required: [name, events, target_url]
      properties:
        name:
          type: string
        events:
          type: array
          items:
            type: string
            enum:
              - execution.started
              - execution.succeeded
              - execution.failed
              - execution.cancelled
              - human_input.required
        target_url:
          type: string
          format: uri
        secret:
          type: string
          description: HMAC-SHA256 signing secret. Stored hashed — shown once.
          nullable: true
        is_active:
          type: boolean
          default: true
    Webhook:
      allOf:
        - $ref: '#/components/schemas/WebhookCreate'
        - type: object
          required: [webhook_id, created_at]
          properties:
            webhook_id:
              type: string
            created_at:
              type: string
              format: date-time

    # ── Audit Log ──────────────────────────────────────────────────────────────
    AuditEvent:
      type: object
      required: [event_id, tenant_id, event_type, resource_id, created_at]
      properties:
        event_id:
          type: string
        tenant_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
          nullable: true
        event_type:
          type: string
          example: workflow.published
        resource_id:
          type: string
        old_value:
          type: object
          nullable: true
        new_value:
          type: object
          nullable: true
        ip_address:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time

    # ── Usage / Billing ────────────────────────────────────────────────────────
    UsageSummary:
      type: object
      required: [tenant_id, period_start, period_end, execution_count, total_cost_usd]
      properties:
        tenant_id:
          type: string
          format: uuid
        period_start:
          type: string
          format: date-time
        period_end:
          type: string
          format: date-time
        execution_count:
          type: integer
        node_exec_count:
          type: integer
        total_input_tokens:
          type: integer
        total_output_tokens:
          type: integer
        total_compute_secs:
          type: number
        total_cost_usd:
          type: number
          format: double

    # ── Team Management ───────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: UNAUTHORIZED
              message: "Bearer token is missing or expired."
              request_id: req_01JRXYZ
    Forbidden:
      description: Authenticated but insufficient permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found (or not visible to this tenant).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  parameters:
    WorkflowId:
      name: workflow_id
      in: path
      required: true
      schema:
        type: string
      description: Workflow identifier (e.g. wf_01JRXYZ).
...
    ScheduleId:
      name: schedule_id
      in: path
      required: true
      schema:
        type: string
    WebhookId:
      name: webhook_id
      in: path
      required: true
      schema:
        type: string
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

# ─────────────────────────────────────────────────────────────────────────────
# Global Security (overridable per-operation)
# ─────────────────────────────────────────────────────────────────────────────
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# Tags
# ─────────────────────────────────────────────────────────────────────────────
...
paths:
...
  /webhooks:
    get:
      operationId: webhooks_list
      tags: [Webhooks]
      summary: List outbound webhooks for the tenant
      responses:
        '200':
          description: Webhook list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Webhook'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: webhooks_create
      tags: [Webhooks]
      summary: Register a new outbound webhook (OWNER only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookCreate'
      responses:
        '201':
          description: Webhook registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /webhooks/{webhook_id}:
    get:
      operationId: webhooks_get
      tags: [Webhooks]
      summary: Get webhook detail
      parameters:
        - $ref: '#/components/parameters/WebhookId'
      responses:
        '200':
          description: Webhook detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      operationId: webhooks_update
      tags: [Webhooks]
      summary: Update webhook configuration
      parameters:
        - $ref: '#/components/parameters/WebhookId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookCreate'
      responses:
        '200':
          description: Webhook updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: webhooks_delete
      tags: [Webhooks]
      summary: Delete a webhook (OWNER only)
      parameters:
        - $ref: '#/components/parameters/WebhookId'
      responses:
        '204':
          description: Webhook deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Webhook Inbound (receive external triggers) ────────────────────────────────
  /webhooks/inbound/{workflow_id}:
    post:
      operationId: webhooks_inbound_trigger
      tags: [Webhooks]
      summary: |
        Inbound webhook — trigger a workflow from an external system.
        Authenticated via HMAC-SHA256 signature in X-Webhook-Signature header.
      security: []
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
        - name: X-Webhook-Signature
          in: header
          required: true
          schema:
            type: string
          description: HMAC-SHA256 of request body, prefixed `sha256=`.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '202':
          description: Webhook received and execution queued.
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Audit Log ─────────────────────────────────────────────────────────────────
  /audit:
    get:
      operationId: audit_list
      tags: [Audit]
      summary: Query tenant audit log (OWNER only — paginated, newest first)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: event_type
          in: query
          schema:
            type: string
          description: Filter by event type (e.g. workflow.published).
        - name: resource_id
          in: query
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: string
            format: uuid
        - name: since
          in: query
          schema:
            type: string
            format: date-time
        - name: until
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Audit events.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditEvent'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ── Usage / Billing ───────────────────────────────────────────────────────────
  /usage:
    get:
      operationId: usage_summary
      tags: [Usage]
      summary: Get aggregated usage for the current billing period
      parameters:
        - name: period_start
          in: query
          schema:
            type: string
            format: date-time
          description: Start of the billing window. Defaults to start of current month.
        - name: period_end
          in: query
          schema:
            type: string
            format: date-time
        - name: granularity
          in: query
          schema:
            type: string
            enum: [hour, day, month]
            default: day
      responses:
        '200':
          description: Usage summary.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/UsageSummary'
                  total_cost_usd:
                    type: number
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ── Team Management ──────────────────────────────────────────────────────────
  # Requires ADMIN or OWNER role for write operations.
...
  /schedules:
    get:
      operationId: schedules_list_all
      tags: [Schedules]
      summary: List all schedules across all workflows for the current tenant
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: is_active
          in: query
          schema:
            type: boolean
          description: Filter by active/inactive status.
        - name: workflow_id
          in: query
          schema:
            type: string
          description: Optionally filter to a single workflow.
      responses:
        '200':
          description: All tenant schedules.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Schedule'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /schedules/{schedule_id}:
    patch:
      operationId: schedules_update
      tags: [Schedules]
      summary: Update a schedule (toggle active, change cron, change timezone)
      parameters:
        - $ref: '#/components/parameters/ScheduleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleCreate'
      responses:
        '200':
          description: Updated schedule.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      operationId: schedules_delete
      tags: [Schedules]
      summary: Delete a schedule
      parameters:
        - $ref: '#/components/parameters/ScheduleId'
      responses:
        '204':
          description: Schedule deleted.
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Integrations — OAuth Connections ─────────────────────────────────────────
  # Third-party credential connections used by api_request, mcp, and connector nodes.
  # OAuth flows redirect to /auth/oauth/{provider} (existing) but are scoped
  # as integration connections rather than user-login connections.
...